refactor: test 파일 data 모듈에 맞게 패키지 생성 및 정리
